media-service/requirements.txt
# media-service Python dependencies
# GStreamer (PyGObject) installed via apt in Dockerfile, not pip.

paho-mqtt>=2.0,<3.0
PyYAML>=6.0,<7.0
psutil>=5.9,<6.0
